requirements.txt
discord.py==1.6.0
tinydb==4.1.1
wavelink==0.9.6
jsonschema==3.2.0
colorlog==4.1.0
pyyaml==5.3.1
schedule==0.6.0
delta==0.4.2
owoify==0.3.1
docopt==0.6.2
